+++ binaries/env/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.33](https://github.com/nativecode-dev/cli-tools/compare/@nofrills/cli-env@0.6.32...@nofrills/cli-env@0.6.33) (2019-09-01)
+
+**Note:** Version bump only for package @nofrills/cli-env
+
+
+
+
+
+## [0.6.32](https://github.com/nativecode-dev/cli-tools/compare/@nofrills/cli-env@0.6.31...@nofrills/cli-env@0.6.32) (2019-09-01)
+
+**Note:** Version bump only for package @nofrills/cli-env
+
+
+
+
+
+## [0.6.31](https://github.com/nativecode-dev/cli-tools/compare/@nofrills/cli-env@0.6.30...@nofrills/cli-env@0.6.31) (2019-09-01)
+
+
+### Bug Fixes
+
+* upgrading packages ([03d279e](https://github.com/nativecode-dev/cli-tools/commit/03d279e))
+
+
+
+
+
+## [0.6.30](https://github.com/nativecode-dev/cli-tools/compare/@nofrills/cli-env@0.6.29...@nofrills/cli-env@0.6.30) (2019-06-27)
+
+**Note:** Version bump only for package @nofrills/cli-env
+
+
+
+
+
 ## [0.6.29](https://github.com/nativecode-dev/cli-tools/compare/@nofrills/cli-env@0.6.24...@nofrills/cli-env@0.6.29) (2019-06-27)
 
 **Note:** Version bump only for package @nofrills/cli-env